requirements-dev.txt
-r requirements.txt
-r requirements-plot.txt

yfinance
pytz
mt5linux @ git+https://github.com/AwesomeTrading/mt5linux.git@master
python-dotenv
python-telegram-bot

# conda install -c conda-forge ta-lib
TA-Lib

pandas_ta
